+++ CHANGELOG.md
@@ -26,6 +26,8 @@ different versioning scheme, following the Haskell community's
 * C++ codegen no longer generates checks for C++11, except for MSVC 2013 workarounds.
 * C++ codegen no longer generates data member initialization that invokes a constructor
   accepting `Comparer` for associative containers.
+* C++ codegen now can generate copy and move constructors with an allocator argument
+  when a custom allocator is used and `--alloc-ctors` is passed to `gbc`.
 
 ### C++ ###